refactor: Make OffroadTagItem
